refactor(client): Fold client_config.h into client.h
